Docs and API improvements
