+++ backend/mirai-console/src/main/kotlin/net/mamoe/mirai/console/command/resolve/CommandCallResolver.kt
@@ -11,8 +11,26 @@ package net.mamoe.mirai.console.command.resolve
 
 import net.mamoe.mirai.console.command.descriptor.ExperimentalCommandDescriptors
 import net.mamoe.mirai.console.command.parse.CommandCall
+import net.mamoe.mirai.console.extensions.CommandCallResolverProvider
+import net.mamoe.mirai.console.internal.extension.GlobalComponentStorage
 
+/**
+ * The resolver converting a [CommandCall] into [ResolvedCommandCall] based on registered []
+ */
+@ExperimentalCommandDescriptors
 public interface CommandCallResolver {
-    @ExperimentalCommandDescriptors
     public fun resolve(call: CommandCall): ResolvedCommandCall?
+
+    public companion object {
+        @JvmStatic
+        @JvmName("resolveCall")
+        public fun CommandCall.resolve(): ResolvedCommandCall? {
+            GlobalComponentStorage.run {
+                CommandCallResolverProvider.useExtensions { provider ->
+                    provider.instance.resolve(this@resolve)?.let { return it }
+                }
+            }
+            return null
+        }
+    }
 }
