+++ backend/mirai-console/src/main/kotlin/net/mamoe/mirai/console/extensions/CommandCallParserProvider.kt
@@ -9,11 +9,15 @@
 
 package net.mamoe.mirai.console.extensions
 
-import net.mamoe.mirai.console.command.resolve.CommandCallResolver
+import net.mamoe.mirai.console.command.descriptor.ExperimentalCommandDescriptors
+import net.mamoe.mirai.console.command.parse.CommandCallParser
 import net.mamoe.mirai.console.extension.AbstractExtensionPoint
 import net.mamoe.mirai.console.extension.InstanceExtension
 
-public interface CommandCallResolverProvider : InstanceExtension<CommandCallResolver> {
-    public companion object ExtensionPoint : AbstractExtensionPoint<CommandCallResolverProvider>(CommandCallResolverProvider::class)
+/**
+ * The resolver for a [CommandCall]
+ */
+@ExperimentalCommandDescriptors
+public interface CommandCallParserProvider : InstanceExtension<CommandCallParser> {
+    public companion object ExtensionPoint : AbstractExtensionPoint<CommandCallParserProvider>(CommandCallParserProvider::class)
 }
-
